requirements.txt
streamlit>=1.30.0
yfinance>=0.2.31
fredapi>=0.5.1
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
plotly>=5.18.0
arch>=6.2.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
